+++ packages/instrumentation-openai/src/instrumentation.ts
@@ -521,7 +521,7 @@ export class OpenAIInstrumentation extends InstrumentationBase<OpenAIInstrumenta
         'OpenAI.Chat.Completions.create stream chunk: %O',
         chunk
       );
-      const idx = chunk.choices[0]?.index ?? 0;
+      const idx = chunk.choices?.[0]?.index ?? 0;
       if (!choices[idx]) {
         choices[idx] = {} as {
           content: string;
@@ -529,7 +529,7 @@ export class OpenAIInstrumentation extends InstrumentationBase<OpenAIInstrumenta
         };
       }
       if (config.captureMessageContent) {
-        const contentPart = chunk.choices[0]?.delta?.content;
+        const contentPart = chunk.choices?.[0]?.delta?.content;
         if (contentPart) {
           if (!choices[idx].content) {
             choices[idx].content = '';
@@ -538,7 +538,7 @@ export class OpenAIInstrumentation extends InstrumentationBase<OpenAIInstrumenta
         }
       }
       // Assume delta.tool_calls, if exists, is an array of length 1.
-      const toolCallPart = chunk.choices[0]?.delta?.tool_calls?.[0];
+      const toolCallPart = chunk.choices?.[0]?.delta?.tool_calls?.[0];
       if (toolCallPart) {
         if (!choices[idx].toolCalls) {
           choices[idx].toolCalls = [];
@@ -575,7 +575,7 @@ export class OpenAIInstrumentation extends InstrumentationBase<OpenAIInstrumenta
         span.setAttribute(ATTR_GEN_AI_RESPONSE_MODEL, model);
       }
       if (!finishReasons[idx]) {
-        const finishReason = chunk.choices[0]?.finish_reason;
+        const finishReason = chunk.choices?.[0]?.finish_reason;
         if (finishReason) {
           finishReasons[idx] = finishReason;
         }