+++ packages/propagation-utils/CHANGELOG.md
@@ -1,6 +1,15 @@
 <!-- markdownlint-disable MD007 MD034 -->
 # Changelog
 
+## [0.31.22](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/propagation-utils-v0.31.21...propagation-utils-v0.31.22) (2026-06-11)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @opentelemetry/contrib-test-utils bumped from ^0.65.0 to ^0.66.0
+
 ## [0.31.21](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/propagation-utils-v0.31.20...propagation-utils-v0.31.21) (2026-05-13)